requirements.txt
absl-py==1.2.0
astor==0.8.1
cycler==0.11.0
fonttools==4.37.4
gast==0.5.3
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
importlib-metadata==5.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
Pillow==9.2.0
protobuf==3.17.3
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zmq==0.0.0
